PKG: Package `simpleio.dll` with `pyparallel`
This commit alters `setup.py` such that `simpleio.dll` will get packaged
with `pyparallel` both when building an `sdist` or a `bdist_wheel`.

For this, we create a "fake subpackage" `parallel.win32`, which
will contain the `simpleio.dll` file (specified in `MANIFEST.in`) from
`src/win32` (specified in `setup.py`: `package_dir` kwarg). That is,
`simpleio.dll` will be installed as `parallel/win32/simpleio.dll` in
the `site-packages` folder.

I have updated `parallelwin32.py` to assume this path accordingly.

I have removed the system-dependent inclusion of `simpleio.dll` in
`setup.py`, meaning it will now be included in all distributions on
all platforms. This is beneficial because it allows to build the
package on a non-Windows platform, but Windows users will still be
able to use it with the included `simpleio.dll`. (Also, I didn't know
how to preserve the old behavior ;))